use world position as id

+++ src/main/java/com/glodblock/github/extendedae/network/packet/SAssemblerMatrixUpdate.java
@@ -15,14 +15,14 @@ import net.neoforged.api.distmarker.OnlyIn;
 
 public class SAssemblerMatrixUpdate implements IMessage {
 
-    private int patternID;
+    private long patternID;
     private Int2ObjectMap<ItemStack> updateMap;
 
     public SAssemblerMatrixUpdate() {
         // NO-OP
     }
 
-    public SAssemblerMatrixUpdate(int id, Int2ObjectMap<ItemStack> updateMap) {
+    public SAssemblerMatrixUpdate(long id, Int2ObjectMap<ItemStack> updateMap) {
         this.patternID = id;
         // deep clone to prevent CME
         this.updateMap = new Int2ObjectOpenHashMap<>(updateMap);
@@ -30,7 +30,7 @@ public class SAssemblerMatrixUpdate implements IMessage {
 
     @Override
     public void toBytes(RegistryFriendlyByteBuf buf) {
-        buf.writeInt(this.patternID);
+        buf.writeLong(this.patternID);
         buf.writeInt(this.updateMap.size());
         for (var entry : this.updateMap.int2ObjectEntrySet()) {
             buf.writeInt(entry.getIntKey());
@@ -40,7 +40,7 @@ public class SAssemblerMatrixUpdate implements IMessage {
 
     @Override
     public void fromBytes(RegistryFriendlyByteBuf buf) {
-        this.patternID = buf.readInt();
+        this.patternID = buf.readLong();
         this.updateMap = new Int2ObjectOpenHashMap<>();
         int size = buf.readInt();
         for (int i = 0; i < size; i ++) {
